+++ lib/types/api.types.ts
@@ -20,7 +20,6 @@ import { AppConfigAPI } from './app.types'
 import { NavigatorAPI } from './navigator.types'
 import { EntryFieldInfo, FieldInfo } from './field.types'
 import { Adapter, KeyValueMap } from 'contentful-management/types'
-import { CMAClient } from './cmaClient.types'
 
 /* User API */
 
@@ -79,10 +78,14 @@ export interface LocationAPI {
 
 /* Parameters API */
 
-export interface ParametersAPI {
-  installation: KeyValueMap
-  instance: KeyValueMap
-  invocation?: SerializedJSONValue
+export interface ParametersAPI<
+  InstallationParameters extends KeyValueMap,
+  InstanceParameters extends KeyValueMap,
+  InvocationParameters extends SerializedJSONValue
+> {
+  installation: InstallationParameters
+  instance: InstanceParameters
+  invocation: InvocationParameters
 }
 
 /* IDs */
@@ -212,7 +215,11 @@ export interface AccessAPI {
 
 type EntryScopedIds = 'field' | 'entry' | 'contentType'
 
-export interface BaseAppSDK {
+export interface BaseAppSDK<
+  InstallationParameters extends KeyValueMap = KeyValueMap,
+  InstanceParameters extends KeyValueMap = KeyValueMap,
+  InvocationParameters extends SerializedJSONValue = SerializedJSONValue
+> {
   /** @deprecated since version 4.0.0 consider using the CMA instead
    * See https://www.contentful.com/developers/docs/extensibility/app-framework/sdk/#using-the-contentful-management-library for more details
    */
@@ -228,7 +235,7 @@ export interface BaseAppSDK {
   /** Methods for displaying notifications. */
   notifier: NotifierAPI
   /** Exposes app configuration parameters */
-  parameters: ParametersAPI
+  parameters: ParametersAPI<InstallationParameters, InstanceParameters, InvocationParameters>
   /** Exposes method to identify app's location */
   location: LocationAPI
   /** Exposes methods for checking user's access level */
@@ -237,17 +244,21 @@ export interface BaseAppSDK {
   ids: Omit<IdsAPI, EntryScopedIds>
   /** Adapter to be injected in contentful-management client */
   cmaAdapter: Adapter
-  /** A CMA Client initialized with default params */
-  cma: CMAClient
 }
 
-export type EditorAppSDK = Omit<BaseAppSDK, 'ids'> &
+export type EditorAppSDK<
+  InstallationParameters extends KeyValueMap = KeyValueMap,
+  InstanceParameters extends KeyValueMap = KeyValueMap
+> = Omit<BaseAppSDK<InstallationParameters, InstanceParameters, never>, 'ids'> &
   SharedEditorSDK & {
     /** A set of IDs for the app */
     ids: Omit<IdsAPI, 'field'>
   }
 
-export type SidebarAppSDK = Omit<BaseAppSDK, 'ids'> &
+export type SidebarAppSDK<
+  InstallationParameters extends KeyValueMap = KeyValueMap,
+  InstanceParameters extends KeyValueMap = KeyValueMap
+> = Omit<BaseAppSDK<InstallationParameters, InstanceParameters, never>, 'ids'> &
   SharedEditorSDK & {
     /** A set of IDs for the app */
     ids: Omit<IdsAPI, 'field'>
@@ -255,7 +266,10 @@ export type SidebarAppSDK = Omit<BaseAppSDK, 'ids'> &
     window: WindowAPI
   }
 
-export type FieldAppSDK = BaseAppSDK &
+export type FieldAppSDK<
+  InstallationParameters extends KeyValueMap = KeyValueMap,
+  InstanceParameters extends KeyValueMap = KeyValueMap
+> = BaseAppSDK<InstallationParameters, InstanceParameters, never> &
   SharedEditorSDK & {
     /** A set of IDs for the app */
     ids: IdsAPI
@@ -265,7 +279,10 @@ export type FieldAppSDK = BaseAppSDK &
     window: WindowAPI
   }
 
-export type DialogAppSDK = Omit<BaseAppSDK, 'ids'> & {
+export type DialogAppSDK<
+  InstallationParameters extends KeyValueMap = KeyValueMap,
+  InvocationParameters extends SerializedJSONValue = SerializedJSONValue
+> = Omit<BaseAppSDK<InstallationParameters, never, InvocationParameters>, 'ids'> & {
   /** A set of IDs for the app */
   ids: Omit<IdsAPI, EntryScopedIds>
   /** Closes the dialog and resolves openCurrentApp promise with data */
@@ -274,29 +291,42 @@ export type DialogAppSDK = Omit<BaseAppSDK, 'ids'> & {
   window: WindowAPI
 }
 
-export type PageAppSDK = Omit<BaseAppSDK, 'ids'> & {
+export type PageAppSDK<InstallationParameters extends KeyValueMap = KeyValueMap> = Omit<
+  BaseAppSDK<InstallationParameters, never, { path: string }>,
+  'ids'
+> & {
   /** A set of IDs actual for the app */
   ids: Omit<IdsAPI, EntryScopedIds>
 }
 
-export type HomeAppSDK = Omit<BaseAppSDK, 'ids'> & {
+export type HomeAppSDK<InstallationParameters extends KeyValueMap = KeyValueMap> = Omit<
+  BaseAppSDK<InstallationParameters, never, never>,
+  'ids'
+> & {
   ids: Omit<IdsAPI, EntryScopedIds>
 }
 
-export type ConfigAppSDK = Omit<BaseAppSDK, 'ids'> & {
+export type ConfigAppSDK<InstallationParameters extends KeyValueMap = KeyValueMap> = Omit<
+  BaseAppSDK<InstallationParameters, never, never>,
+  'ids'
+> & {
   /** A set of IDs actual for the app */
   ids: Omit<IdsAPI, EntryScopedIds | 'extension' | 'app'> & { app: string }
   app: AppConfigAPI
 }
 
-export type KnownAppSDK =
-  | FieldAppSDK
-  | SidebarAppSDK
-  | DialogAppSDK
-  | EditorAppSDK
-  | PageAppSDK
-  | ConfigAppSDK
-  | HomeAppSDK
+export type KnownAppSDK<
+  InstallationParameters extends KeyValueMap = KeyValueMap,
+  InstanceParameters extends KeyValueMap = KeyValueMap,
+  InvocationParameters extends SerializedJSONValue = SerializedJSONValue
+> =
+  | FieldAppSDK<InstallationParameters, InstanceParameters>
+  | SidebarAppSDK<InstallationParameters, InstanceParameters>
+  | DialogAppSDK<InstallationParameters, InvocationParameters>
+  | EditorAppSDK<InstallationParameters, InstanceParameters>
+  | PageAppSDK<InstallationParameters>
+  | ConfigAppSDK<InstallationParameters>
+  | HomeAppSDK<InstallationParameters>
 
 /** @deprecated consider using {@link BaseAppSDK} */
 export type BaseExtensionSDK = BaseAppSDK
@@ -339,7 +369,7 @@ export interface Locations {
 export interface ConnectMessage {
   id: string
   location: Location[keyof Location]
-  parameters: ParametersAPI
+  parameters: ParametersAPI<KeyValueMap, KeyValueMap, never>
   locales: LocalesAPI
   user: UserAPI
   initialContentTypes: ContentType[]